Refactor: structure(run_bc_from_scratch)
